fix: rename table file, dev MuiDecorator
