+++ src/lib/table/index.ts
@@ -1,3 +1,3 @@
-export { Table } from './table';
+export { Table } from './table-component';
 export { TableVirtualized } from './table-virtualized';
 export type { TableColumn, TableColumnDef } from './types';
